Remove trailing slash from filename
Avoiding a bug where if the server directly opens the TW file instead of its directory (to avoid the "view the folder listing, and click on your wiki html file" step), a trailing slash is added after .html and Ruby can no longer save the file:

[2021-12-18 15:44:10] ERROR Errno::ENOTDIR: Not a directory @ rb_sysopen - test/test.html/

For some reason it doesn't want to work when TW is saved as index.html, but I feel like this workaround is sufficient. I've also expanded on the usage notes.